updated delete all entities function

+++ src/functions/delete-all-entities.function.ts
@@ -6,10 +6,15 @@ import { Order } from "../entity/order.entity";
 import { ProductCategory } from "../entity/product-category.entity";
 import { Product } from "../entity/product.entity";
 
+let notFirstCall = false;
+
 export async function deleteAllEntities() {
     const tableNames = [Address.tableName, Customer.tableName, Order.tableName, OrderItem.tableName, Product.tableName, ProductCategory.tableName];
     for (const tableName of tableNames) {
-        await DB.query(`DELETE from [${tableName}];`)
-        await DB.query(`DBCC CHECKIDENT ([${tableName}], RESEED, 0);`);
+        if (notFirstCall) {
+            await DB.query(`DELETE from [${tableName}];`)
+            await DB.query(`DBCC CHECKIDENT ([${tableName}], RESEED, 0);`);
+        }
     }
+    notFirstCall = true;
 }
